+++ rororo/openapi/data.py
@@ -1,3 +1,12 @@
+"""
+===================
+rororo.openapi.data
+===================
+
+Provide structures for OpenAPI data.
+
+"""
+
 from typing import Any
 
 import attr
@@ -26,6 +35,43 @@ class OpenAPIParameters:
 
 @attr.dataclass(frozen=True, slots=True)
 class OpenAPIContext:
+    """All data associated with current request to OpenAPI handler.
+
+    Contains only valid parameters, security data & request body data.
+    Example bellow illustrates how to work with context data,
+
+    .. code-block:: python
+
+        from rororo import get_openapi_context
+        from rororo.openapi.exceptions import InvalidCredentials
+
+
+        async def create_user(request: web.Request) -> web.Response:
+            context = get_openapi_context(request)
+
+            # Authenticate current user (accessing security data)
+            if not authenticate(api_key=context.security["apiKey"]):
+                raise InvalidCredentials()
+
+            # Add new user (accessing request body data)
+            async with request.config_dict["db"].acquire() as conn:
+                user = await create_user(
+                    conn,
+                    email=context.data["email"],
+                    password=context.data["password"],
+                )
+
+            # Return response due to query string param
+            # (accessing parameters data)
+            if context.parameters.query["login"]:
+                return web.json_response(
+                    request.app.router["login"].url_for()
+                )
+
+            return web.json_response(user.to_api_dict())
+
+    """
+
     #: Request instance
     request: web.Request